Temorary spec for generator
You can run generator spec by doing:

    GENERATOR_SPEC=true rake spec:ative_record_41

We are doing this because we don't want to make API calls in the
spec. At the same time, we aldo don't want to use fakeweb or something
like that because it causes "you never realize until you go production"
 situation. We had it before and shipped a bug - we are not sure what is
a better solution, but at long as I'm aware of this I'll take care of it.