add :use_page_one_param option
When passed as an option to the paginate() helper, links pointing
to page one will always include the `page=1` param, rather than the
default behavior of leaving the `page` parameter off altogether. This
is useful for certain implementations where page 1 is not the
default page when the `page` parameter is not present.