Add possibility to set `total_count` for a scope
Counting can be slow with large DB sets. Especially in PostgreSQL,
because it returns a precise result and therefore it needs to read all
the resulting rows to verify that they exists.

But sometimes it is better to have a fast but not so precise result. In
that case we can tell `kaminari` what the approximate total count is, so
it doesn't need to calculate it with SQL `count`.