add #repaginate_out_of_range method
This allows to define a custom behavior when #page is called with a value out of
range.

For example, when setting
```ruby
  # given there are 100 articles
  Article.page(10).per(25).all
  # => []

  Article.when_page_is_out_of_range_set :first
  Article.page(10).per(25).all
  # => # the first 25 articles

  Article.when_page_is_out_of_range_set :last
  Article.page(10).per(25).all
  # => # the last 25 articles
```

This behavior is useful when you are deleting resources from a list and you are
on the last page. If the last page has only one resource, after deleting it you
can redirect the user to the previous (last) page or force the first page.