Remove the need to call #presence to show a fall-back HTML
Right now, the #paginate method doesn't show anything when there's only one
page. While this is expected, some times there's a case where showing a fall
back message is appropriate. In order to do so, a subsequent #presence call is
needed because the #paginate method always returns a string. This commit adds a
 #presence call to the #render method, so it'll be a little cleaner to show a
fall-back message.

  related: https://github.com/kaminari/kaminari/pull/785