Optimize page_entries_info a bit by reusing local variables
The entry_name local variable uses collection.size to pluralize the
name. This value is calculated in all cases. The size of the collection
is reused later to calculate the `to` local variable (the final entry on
the page).

By reusing this calculated value, we cut down on database round trips.
When the collection query is very expensive, running it twice can cause
a large page slowdown.