Add notes about providing an order
We got stung recently by a bug where a developer assumed that `User.per(n).page(n)` would order by `users.id`. Instead, it generates an un-ordered query, which leads to undefined ordering at the database layer.

While it's not strictly a responsibility of Kaminari to enforce this, I think reminding users in the docs that they still need to order their queries would be helpful for future developers.