Use `offset_value` for calculating `total_count`
This fixes a bug where `total_count` returns an incorrect value in the end of list on `loaded` Relation when `padding` is specified.

When padding is given, it internally calls `offset` directly. Thus, in order to cover that case, we need to reference `offset_value` instead of `page * limit` for calculating `total_count`.

fixes #1055 closes #1085